traffic/requirements.txt
psycopg2-binary==2.9.3
requests==2.27.1
numpy==1.21.3
matplotlib
pandas
